tqdm logging
